Fix some compile warnings
Mostly unsafe type conversions, and replacing deprecated APIs.

Still can't remove the not-clean-build flags because of bug in qtbase.

(cherry picked from commit 5729b957a9db727345d6b5f3493791f05f922a88)

+++ src/plugins/geometryloaders/default/stlgeometryloader.cpp
@@ -90,7 +90,7 @@ bool StlGeometryLoader::loadAscii(QIODevice *ioDev)
                 const float y = tokens.floatAt(2);
                 const float z = tokens.floatAt(3);
                 m_points.push_back(QVector3D(x, y, z));
-                m_indices.push_back(m_indices.size());
+                m_indices.push_back(uint(m_indices.size()));
             }
         }
     }
